chore: migrate to Lerna, add TypeScript
