+++ README.md
@@ -1,5 +1,7 @@
 # Fullstack Sandbox
 
+[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
+
 ## Frontend
 
 ## Backend
