+++ lerna.json
@@ -1,0 +1,6 @@
+{
+  "version": "0.0.0",
+  "packages": ["packages/*"],
+  "npmClient": "yarn",
+  "useWorkspaces": true
+}
