+++ packages/backend/src/index.js
@@ -1,0 +1,23 @@
+const express = require('express')
+const { ApolloServer, gql } = require('apollo-server-express')
+
+const typeDefs = gql`
+  type Query {
+    hello: String
+  }
+`
+
+const resolvers = {
+  Query: {
+    hello: () => 'Hello, World!',
+  },
+}
+
+const server = new ApolloServer({ typeDefs, resolvers })
+
+const app = express()
+server.applyMiddleware({ app })
+
+app.listen({ port: 4000 }, () =>
+  console.log(`🚀 Server ready at http://localhost:4000${server.graphqlPath}`)
+)
