+++ packages/frontend/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "frontend",
+  "name": "@fullstack-sandbox/frontend",
   "private": true,
-  "version": "1.0.0",
+  "version": "0.0.0",
   "description": "A simple NextJS App",
   "keywords": [
     "react",
@@ -14,16 +14,12 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "static:typecheck": "tsc"
   },
   "dependencies": {
     "next": "9.0.4",
     "react": "16.9.0",
     "react-dom": "16.9.0"
-  },
-  "devDependencies": {
-    "@types/node": "^12.7.2",
-    "@types/react": "^16.9.2",
-    "typescript": "^3.5.3"
   }
 }
