+++ packages/frontend/pages/index.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import Head from "next/head";
+import React from 'react'
+import Head from 'next/head'
 
 export default function Home() {
   return (
@@ -12,5 +12,5 @@ export default function Home() {
         <h1>Welcome to Next.js!</h1>
       </div>
     </div>
-  );
+  )
 }
