+++ packages/frontend/static/favicon.ico
